+++ test/blutwurst/value_generators_test.clj
@@ -1,4 +1,5 @@
 (ns blutwurst.value-generators-test
+  (:import (java.lang Math))
   (:require [clojure.test :refer :all]
             [clojure.pprint :refer :all]
             [blutwurst.value-generators :as vg]))
@@ -23,7 +24,42 @@
    }
   ])
 
-(deftest random-string-test
+(deftest random-value-test
+ (testing "Basic random number generator respects size of smallint."
+  (let [generator-fn (->> vg/value-generation-strategies
+                         (filter #(= "Random Integer Generator" (:name %)))
+                         first
+                         :generator)]
+    (dotimes [iter 100]
+     (let [value (generator-fn {:name "foo" :type "TINYINT"})]
+       (is (and (<= value 255)
+                (>= value 0))
+           "SMALLINT must be in the range of values for one byte.")
+    ))
+
+    ; TODO: these should all allow for negative values.
+    (dotimes [iter 100]
+     (let [value (generator-fn {:name "foo" :type "SMALLINT"})]
+       (is (and (<= value (- (Math/pow 2 15) 1))
+                (>= value 0))
+           "INTEGER must be in the range of values for 4 bytes.")
+    ))
+
+    (dotimes [iter 100]
+     (let [value (generator-fn {:name "foo" :type "BIGINT"})]
+       (is (and (<= value (- (Math/pow 2 63) 1))
+                (>= value 0))
+           "INTEGER must be in the range of values for 4 bytes.")
+    ))
+
+    (dotimes [iter 100]
+     (let [value (generator-fn {:name "foo" :type "INTEGER"})]
+       (is (and (<= value (- (Math/pow 2 32) 1))
+                (>= value 0))
+           "INTEGER must be in the range of values for 4 bytes.")
+    ))
+   ))
+
  (testing "Basic random string generation."
    (let [value (vg/random-string 10)]
     (is (< (count value) 10)))